+++ packages/perseus-core/src/parse-perseus-json/regression-tests/item-data/numeric-input-answer-with-simplify-true.ts
@@ -1,0 +1,50 @@
+// WARNING: Do not change or delete this file! If you do, Perseus might become
+// unable to parse the current data format, which will break clients.
+// If you need to add more regression tests, add a new file to this directory.
+export default {
+    answerArea: {
+        calculator: false,
+    },
+    hints: [
+        {
+            content: "$= \\dfrac{9 \\times 1}{2 \\times 4}$",
+        },
+        {
+            content: "$= \\dfrac{9}{8}$",
+        },
+    ],
+    itemDataVersion: {
+        major: 0,
+        minor: 1,
+    },
+    question: {
+        content:
+            "**$ \\dfrac{9}{2} \\times \\dfrac{1}{4} $**\n\n[[☃ numeric-input 1]]",
+        widgets: {
+            "numeric-input 1": {
+                alignment: "default",
+                graded: true,
+                options: {
+                    answers: [
+                        {
+                            maxError: 0,
+                            message: "",
+                            simplify: true,
+                            status: "correct",
+                            strict: false,
+                            value: 1.125,
+                        },
+                    ],
+                    coefficient: false,
+                    labelText: "",
+                    size: "normal",
+                },
+                type: "numeric-input",
+                version: {
+                    major: 0,
+                    minor: 0,
+                },
+            },
+        },
+    },
+};