+++ packages/perseus-core/src/parse-perseus-json/regression-tests/item-data/passage-missing-footnotes-and-title.ts
@@ -1,0 +1,49 @@
+// WARNING: Do not change or delete this file! If you do, Perseus might become
+// unable to parse the current data format, which will break clients.
+// If you need to add more regression tests, add a new file to this directory.
+export default {
+    answerArea: {
+        calculator: false,
+        options: {
+            content: "",
+            images: {},
+            widgets: {},
+        },
+        type: "multiple",
+    },
+    hints: [],
+    itemDataVersion: {
+        major: 0,
+        minor: 1,
+    },
+    question: {
+        content: "\n\n[[☃ passage 1]]\n\n[[☃ passage-ref 1]]",
+        images: {},
+        widgets: {
+            "passage 1": {
+                graded: true,
+                options: {
+                    passageText: "a",
+                    showLineNumbers: true,
+                },
+                type: "passage",
+                version: {
+                    major: 0,
+                    minor: 0,
+                },
+            },
+            "passage-ref 1": {
+                graded: true,
+                options: {
+                    passageNumber: 1,
+                    referenceNumber: 1,
+                },
+                type: "passage-ref",
+                version: {
+                    major: 0,
+                    minor: 0,
+                },
+            },
+        },
+    },
+};