+++ packages/perseus-core/src/parse-perseus-json/regression-tests/user-input-data/free-response.ts
@@ -1,4 +1,4 @@
 // WARNING: Do not change or delete this file! If you do, Perseus might become
-// unable to parse the current user input format, which will break clients.
+// unable to parse the current data format, which will break clients.
 // If you need to add more regression tests, add a new file to this directory.
 export default {"free-response 1": {currentValue: "hello"}};