+++ packages/perseus/src/widgets/explanation/explanation.tsx
@@ -1,6 +1,7 @@
 import {linterContextDefault} from "@khanacademy/perseus-linter";
 import Button from "@khanacademy/wonder-blocks-button";
 import {Id, View} from "@khanacademy/wonder-blocks-core";
+import {font} from "@khanacademy/wonder-blocks-tokens";
 import caretDown from "@phosphor-icons/core/regular/caret-down.svg";
 import caretUp from "@phosphor-icons/core/regular/caret-up.svg";
 import {StyleSheet} from "aphrodite";
@@ -8,6 +9,7 @@ import * as React from "react";
 
 import {PerseusI18nContext} from "../../components/i18n-context";
 import Renderer from "../../renderer";
+import UserInputManager from "../../user-input-manager";
 import {getPromptJSON as _getPromptJSON} from "../../widget-ai-utils/explanation/explanation-ai-utils";
 
 import type {Widget, WidgetExports, WidgetProps} from "../../types";
@@ -97,7 +99,7 @@ class Explanation extends React.Component<Props, State> implements Widget {
         };
 
         const labelStyle = {
-            fontSize: "18px",
+            fontSize: font.heading.size.medium,
             lineHeight: "inherit",
             "text-align": "left",
             // The following property adjusts the large space between the button text and the caret icon.
@@ -142,13 +144,37 @@ class Explanation extends React.Component<Props, State> implements Widget {
                             testId="content-container"
                         >
                             <View style={styles.contentWrapper}>
-                                <Renderer
-                                    apiOptions={this.props.apiOptions}
-                                    content={this.props.explanation}
+                                <UserInputManager
                                     widgets={this.props.widgets}
-                                    linterContext={this.props.linterContext}
-                                    strings={this.context.strings}
-                                />
+                                    problemNum={0}
+                                >
+                                    {({
+                                        userInput,
+                                        handleUserInput,
+                                        initializeUserInput,
+                                    }) => {
+                                        return (
+                                            <Renderer
+                                                apiOptions={
+                                                    this.props.apiOptions
+                                                }
+                                                content={this.props.explanation}
+                                                widgets={this.props.widgets}
+                                                linterContext={
+                                                    this.props.linterContext
+                                                }
+                                                strings={this.context.strings}
+                                                userInput={userInput}
+                                                handleUserInput={
+                                                    handleUserInput
+                                                }
+                                                initializeUserInput={
+                                                    initializeUserInput
+                                                }
+                                            />
+                                        );
+                                    }}
+                                </UserInputManager>
                             </View>
                         </View>
                     </>