+++ packages/perseus/src/widgets/image/__docs__/image-initial-state-regression.stories.tsx
@@ -1,17 +1,7 @@
-import {
-    generateTestPerseusRenderer,
-    generateImageOptions,
-    generateImageWidget,
-} from "@khanacademy/perseus-core";
 import * as React from "react";
 
-import {themeModes} from "../../../../../../.storybook/modes";
-import {getFeatureFlags} from "../../../../../../testing/feature-flags-util";
-import {ApiOptions} from "../../../perseus-api";
-import Renderer from "../../../renderer";
-import {mockStrings} from "../../../strings";
-import UserInputManager from "../../../user-input-manager";
 import {getWidget} from "../../../widgets";
+import {imageRendererDecorator} from "../../__testutils__/image-renderer-decorator";
 import {
     mobileDecorator,
     articleDecorator,
@@ -20,7 +10,6 @@ import {
 } from "../../__testutils__/story-decorators";
 import {earthMoonImage, frescoImage, monasteryImage} from "../utils";
 
-import type {PerseusRenderer} from "@khanacademy/perseus-core";
 import type {Meta, StoryObj} from "@storybook/react-vite";
 
 const ImageWidget = getWidget("image")!;
@@ -33,26 +22,6 @@ const frescsoLongDescription =
     "In the apse, or semicircular recess, The Offer of the Casa Madre to Victory (L’Offerta della Casa Madre alla Vittoria) fresco recalls medieval apse decorative schemes with Christ surrounded by saints to whom the Church is dedicated. Santagata replaced Mary with a triumphant and wingless figure representing Victory, and he replaced saints with sentries. The charismatic wounded veteran Carlo Delcroix, who became the AMNIG president, is depicted presenting a model of the Casa Madre to Victory (not unlike the medieval patron Enrico Scrovegni, who offered the Arena chapel he commissioned to the Virgin Mary).\n\nThis image has some stuff in it. *Here is some italic text.* **Here is some bold text.**";
 
 const articleContent = `But in other cases, an object may experience a centripetal force for an extended time and complete *repeated* revolutions. An example of this type of motion is an astronomical object in **orbit**.\n\n[[☃ image 1]]\n\nLet's explore some of the language and relationships involved in orbital motion.`;
-
-// Breaking this out instead of using it globally, so that the
-// right-to-left story can wrap the ImageQuestionRenderer with the
-// right-to-left wrapper.
-const rendererDecorator = (_, {args, parameters}) => {
-    return (
-        <ImageQuestionRenderer
-            question={generateTestPerseusRenderer({
-                content: parameters?.content ?? "[[☃ image 1]]",
-                widgets: {
-                    "image 1": generateImageWidget({
-                        options: generateImageOptions({
-                            ...args,
-                        }),
-                    }),
-                },
-            })}
-        />
-    );
-};
 
 const meta: Meta<typeof ImageWidget> = {
     title: "Widgets/Image/Visual Regression Tests",
@@ -70,42 +39,17 @@ const meta: Meta<typeof ImageWidget> = {
 export default meta;
 
 export const Image: Story = {
-    decorators: [rendererDecorator],
-    args: {
-        backgroundImage: earthMoonImage,
-    },
-};
-
-export const ImageWithAll: Story = {
-    decorators: [rendererDecorator],
+    decorators: [imageRendererDecorator],
     args: {
         backgroundImage: earthMoonImage,
         alt: "Earth and Moon",
-        longDescription:
-            "This is a *very* long description of the earth and moon.",
         title: "Earth and Moon",
         caption: earthMoonImageCaption,
     },
 };
 
-export const ImageWithAlt: Story = {
-    decorators: [rendererDecorator],
-    args: {
-        backgroundImage: earthMoonImage,
-        alt: "Earth and Moon",
-    },
-};
-
-export const ImageWithCaption: Story = {
-    decorators: [rendererDecorator],
-    args: {
-        backgroundImage: earthMoonImage,
-        caption: earthMoonImageCaption,
-    },
-};
-
 export const ImageWithCaptionAndLongDescription: Story = {
-    decorators: [rendererDecorator],
+    decorators: [imageRendererDecorator],
     args: {
         backgroundImage: earthMoonImage,
         caption: earthMoonImageCaption,
@@ -115,7 +59,7 @@ export const ImageWithCaptionAndLongDescription: Story = {
 };
 
 export const ImageWithLongDescription: Story = {
-    decorators: [rendererDecorator],
+    decorators: [imageRendererDecorator],
     args: {
         backgroundImage: earthMoonImage,
         longDescription:
@@ -124,42 +68,25 @@ export const ImageWithLongDescription: Story = {
 };
 
 export const ImageWithTitle: Story = {
-    decorators: [rendererDecorator],
+    decorators: [imageRendererDecorator],
     args: {
         backgroundImage: earthMoonImage,
         title: "Earth and Moon",
     },
 };
 
-export const MobileImageAll: Story = {
-    decorators: [rendererDecorator, mobileDecorator],
+export const MobileImage: Story = {
+    decorators: [imageRendererDecorator, mobileDecorator],
     args: {
         backgroundImage: earthMoonImage,
         caption: earthMoonImageCaption,
-        longDescription:
-            "This is a *very* long description of the earth and moon.",
         title: "Earth and Moon",
-    },
-};
-
-export const MobileImageWithAlt: Story = {
-    decorators: [rendererDecorator, mobileDecorator],
-    args: {
-        backgroundImage: earthMoonImage,
         alt: "Earth and Moon",
     },
 };
 
-export const MobileImageWithCaption: Story = {
-    decorators: [rendererDecorator, mobileDecorator],
-    args: {
-        backgroundImage: earthMoonImage,
-        caption: earthMoonImageCaption,
-    },
-};
-
 export const MobileImageWithCaptionAndLongDescription: Story = {
-    decorators: [rendererDecorator, mobileDecorator],
+    decorators: [imageRendererDecorator, mobileDecorator],
     args: {
         backgroundImage: earthMoonImage,
         caption: earthMoonImageCaption,
@@ -169,7 +96,7 @@ export const MobileImageWithCaptionAndLongDescription: Story = {
 };
 
 export const MobileImageWithLongDescription: Story = {
-    decorators: [rendererDecorator, mobileDecorator],
+    decorators: [imageRendererDecorator, mobileDecorator],
     args: {
         backgroundImage: earthMoonImage,
         longDescription:
@@ -178,7 +105,7 @@ export const MobileImageWithLongDescription: Story = {
 };
 
 export const MobileImageWithTitle: Story = {
-    decorators: [rendererDecorator, mobileDecorator],
+    decorators: [imageRendererDecorator, mobileDecorator],
     args: {
         backgroundImage: earthMoonImage,
         title: "Earth and Moon",
@@ -186,7 +113,7 @@ export const MobileImageWithTitle: Story = {
 };
 
 export const PortraitImage: Story = {
-    decorators: [rendererDecorator],
+    decorators: [imageRendererDecorator],
     args: {
         backgroundImage: monasteryImage,
         caption:
@@ -197,8 +124,8 @@ export const PortraitImage: Story = {
     },
 };
 
-export const WithinArticleDesktop: Story = {
-    decorators: [rendererDecorator, articleDecorator],
+export const WithinArticle: Story = {
+    decorators: [imageRendererDecorator, articleDecorator],
     parameters: {
         content: articleContent,
     },
@@ -208,8 +135,8 @@ export const WithinArticleDesktop: Story = {
     },
 };
 
-export const WithinArticleDesktopCaptionAndTitle: Story = {
-    decorators: [rendererDecorator, articleDecorator],
+export const WithinArticleWithCaptionAndTitle: Story = {
+    decorators: [imageRendererDecorator, articleDecorator],
     parameters: {
         content: articleContent,
     },
@@ -222,7 +149,7 @@ export const WithinArticleDesktopCaptionAndTitle: Story = {
 };
 
 export const WithinArticleMobile: Story = {
-    decorators: [rendererDecorator, mobileArticleDecorator],
+    decorators: [imageRendererDecorator, mobileArticleDecorator],
     parameters: {
         content: articleContent,
     },
@@ -233,7 +160,7 @@ export const WithinArticleMobile: Story = {
 };
 
 export const WithinArticleMobileCaptionAndTitle: Story = {
-    decorators: [rendererDecorator, mobileArticleDecorator],
+    decorators: [imageRendererDecorator, mobileArticleDecorator],
     parameters: {
         content: articleContent,
     },
@@ -246,7 +173,7 @@ export const WithinArticleMobileCaptionAndTitle: Story = {
 };
 
 export const RightToLeftImage: Story = {
-    decorators: [rendererDecorator, rtlDecorator],
+    decorators: [imageRendererDecorator, rtlDecorator],
     args: {
         alt: "Fresco of some people",
         title: "The Offer of the Casa Madre to Victory, 1932",
@@ -259,7 +186,7 @@ export const RightToLeftImage: Story = {
 
 export const RightToLeftImageMobile: Story = {
     decorators: [
-        rendererDecorator,
+        imageRendererDecorator,
         mobileDecorator,
         (Story) => (
             <div style={{direction: "rtl"}}>
@@ -277,25 +204,28 @@ export const RightToLeftImageMobile: Story = {
     },
 };
 
-function ImageQuestionRenderer(props: {question: PerseusRenderer}) {
-    const {question} = props;
-    return (
-        <UserInputManager widgets={question.widgets} problemNum={0}>
-            {({userInput, handleUserInput, initializeUserInput}) => (
-                <Renderer
-                    userInput={userInput}
-                    handleUserInput={handleUserInput}
-                    initializeUserInput={initializeUserInput}
-                    strings={mockStrings}
-                    content={question.content}
-                    widgets={question.widgets}
-                    images={question.images}
-                    apiOptions={{
-                        ...ApiOptions.defaults,
-                        flags: getFeatureFlags({"image-widget-upgrade": true}),
-                    }}
-                />
-            )}
-        </UserInputManager>
-    );
-}
+export const DecorativeImage: Story = {
+    decorators: [imageRendererDecorator],
+    args: {
+        backgroundImage: earthMoonImage,
+        alt: "Earth and Moon",
+        title: "Earth and Moon",
+        caption: earthMoonImageCaption,
+        longDescription:
+            "This is a *very* long description of the earth and moon.",
+        decorative: true,
+    },
+};
+
+export const DecorativeImageMobile: Story = {
+    decorators: [imageRendererDecorator, mobileDecorator],
+    args: {
+        backgroundImage: earthMoonImage,
+        alt: "Earth and Moon",
+        title: "Earth and Moon",
+        caption: earthMoonImageCaption,
+        longDescription:
+            "This is a *very* long description of the earth and moon.",
+        decorative: true,
+    },
+};